+++ packages/awino-ui/next.config.js
@@ -1,13 +1,23 @@
 const { i18n } = require('./next-i18next.config');
 // const { PHASE_DEVELOPMENT_SERVER } = require('next/constants')
 
-module.exports = async (phase, { defaultConfig }) => {
+module.exports = async (/* phase, { defaultConfig } */) => {
   /**
     * @type {import('./src/types').NextAppConfig}
    */
   let nextConfig = {
     i18n,
-    /* config options here */
+    reactStrictMode: true,
+    // images: {
+    //   domains: ['storage.googleapis.com'],
+    //   deviceSizes: [640, 750, 828, 1080, 1200, 1920, 2048, 3840],
+    //   imageSizes: [16, 32, 48, 64, 96, 128, 256, 384]
+    // },
+    serverRuntimeConfig: {
+    },
+    publicRuntimeConfig: {
+      baseDomain: process.env.NEXT_PUBLIC_BASE_DOMAIN,
+    },
   }
 
   return nextConfig;