+++ packages/awino-ui/src/app/constants.ts
@@ -1,12 +1,11 @@
-// import getConfig from './env';
-// import { Language } from '@/typings/app';
-
 import type { Language } from '@/types/app';
 
-// const {
-//   serverRuntimeConfig,
-//   publicRuntimeConfig: { baseDomain, cmsEndpoint, googleRecaptchaSiteKey, subscribeEndpoint },
-// } = getConfig();
+import getConfig from './env';
+
+const {
+  // serverRuntimeConfig,
+  publicRuntimeConfig: { baseDomain },
+} = getConfig();
 
 export const SUPPORTED_LANGUAGES: Language[] = ['en', 'de'];
 
@@ -22,7 +21,7 @@ export const TABLE_ROWS_PER_PAGE = TABLE_ROWS_PER_PAGE_OPTIONS[0];
 
 // export const DEFAULT_DATE_TIME_FORMAT = 'yyyy-MM-dd HH:mm';
 
-// export const BASE_DOMAIN = baseDomain;
+export const BASE_DOMAIN = baseDomain;
 
 // export const CMS_ENDPOINT = serverRuntimeConfig?.cmsEndpoint || cmsEndpoint;