+++ packages/awino-ui/src/app/env.ts
@@ -1,0 +1,10 @@
+import getNextConfig from 'next/config';
+
+import { NextAppConfig } from '@/types/app';
+
+const getConfig = (): NextAppConfig => {
+  const { serverRuntimeConfig = {}, publicRuntimeConfig = {} } = getNextConfig() || {};
+  return { serverRuntimeConfig, publicRuntimeConfig };
+};
+
+export default getConfig;